# E-Commerce/openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "OpenAPI definition",
		"version": "v0"
	},
	"servers": [
		{
			"url": "http://localhost:8080",
			"description": "Generated server url"
		}
	],
	"paths": {
		"/api/auth/seller-register": {
			"post": {
				"tags": [
					"auth-controller"
				],
				"operationId": "sellerRegisterHandler",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Users"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/auth/register": {
			"post": {
				"tags": [
					"auth-controller"
				],
				"operationId": "registerHandler",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Users"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"tags": [
					"auth-controller"
				],
				"operationId": "loginHandler",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Users"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"GrantedAuthority": {
				"type": "object",
				"properties": {
					"authority": {
						"type": "string"
					}
				}
			},
			"Users": {
				"type": "object",
				"properties": {
					"userId": {
						"type": "integer",
						"format": "int32"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"deleted": {
						"type": "string"
					},
					"isUserSeller": {
						"type": "string"
					},
					"createdBy": {
						"type": "integer",
						"format": "int32"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedBy": {
						"type": "integer",
						"format": "int32"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"macAddress": {
						"type": "string"
					},
					"ipAddress": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"countryCode": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"countryId": {
						"type": "integer",
						"format": "int32"
					},
					"businessName": {
						"type": "string"
					},
					"isAdmin": {
						"type": "string"
					},
					"authorities": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/GrantedAuthority"
						}
					}
				}
			},
			"AuthResponse": {
				"type": "object",
				"properties": {
					"token": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				}
			},
...
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"scheme": "bearer"
			}
		}
	}
}